Settings to get user data from LDAP
This allows to retrieve any part(s) of user data from a LDAP directory (like AD controller), so that the data would be retrieved automatically for each user. When any part is configured to be taken from LDAP, its value defines the name of LDAP object's attribute which holds the requested data.

For this configuration to work, new LDAP connection settings section must be filled properly.