Option to disable "allow this website to open a program" warning in IE
The option uses WarnOnOpen setting located under user's or computer's
SOFTWARE\Microsoft\Internet Explorer\ProtocolExecute, as described in
https://blogs.msdn.microsoft.com/ieinternals/2011/07/13/understanding-protocols/

This is a security-sensitive setting. As mentioned in the aforementioned
article, "for security reasons suppressing this prompt is generally
discouraged".

Since there's no way to enforce a multi-value setting outside of GPO
registry key, it's possible for a user (or a program) to change this
value after being set by GPO. Thus, the setting is rather to set a
"default" rather than to guarantee behaviour.